DEPR: Enforce empty Series returning object dtype
